backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
easyocr
pymongo
pydantic
python-dotenv
motor
pillow
numpy
python-doctr[torch]
paddlepaddle
paddleocr
passlib
